backend/requirements.txt
fastapi
uvicorn
pydantic
httpx

google-generativeai
python-dotenv
